% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_alphas.R
\name{alphas_list_to_data_frame}
\alias{alphas_list_to_data_frame}
\title{Cast a list of alphas as data frame}
\usage{
alphas_list_to_data_frame(alphas_grid)
}
\arguments{
\item{alphas_grid}{A list of length 3 parameter tuples as returned by
\code{\link[=list_sampled_alphas]{list_sampled_alphas()}} or \code{\link[=list_edge_alphas]{list_edge_alphas()}}}
}
\value{
A data frame with three numeric columns called \code{alpha},
\code{alpha_seasonal}, and \code{alpha_seasonal_decay}
}
\description{
A helper function that casts a list of parameter tuples (as used for
\code{\link[=learn_weights]{learn_weights()}}) as a named data frame. This is useful to, for example,
plot the parameters via \code{ggplot2}.
}
\examples{
alphas_list_to_data_frame(list_sampled_alphas(n_target = 10L))

}
\seealso{
\code{\link[=list_sampled_alphas]{list_sampled_alphas()}}, \code{\link[=list_edge_alphas]{list_edge_alphas()}}
}
