% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.threedx}
\alias{autoplot.threedx}
\title{Autoplot method for \code{threedx} objects}
\usage{
\method{autoplot}{threedx}(object, ..., date = NULL, show_params = TRUE)
}
\arguments{
\item{object}{Fitted model object of class \code{threedx} returned by
\code{\link[=learn_weights]{learn_weights()}}}

\item{...}{ignored}

\item{date}{Optional additional vector with dates in format that can be cast
to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create x-axis}

\item{show_params}{Logical; if \code{TRUE} (default) then fitted parameters will
be displayed using \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}
}
\description{
Use \code{ggplot2} to visualize the fitted values and weights of a fitted model
of class \code{threedx}
}
