% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.threedx_paths}
\alias{autoplot.threedx_paths}
\title{Autoplot method for \code{threedx_paths} objects}
\usage{
autoplot.threedx_paths(
  object,
  ...,
  method = c("forecast", "paths")[1],
  date = NULL,
  date_future = NULL,
  show_params = TRUE,
  n = 5,
  alpha = 0.75
)
}
\arguments{
\item{object}{An object of class \code{threedx_paths} as returned by
\verb{[predict.threedx()]}}

\item{...}{ignored}

\item{method}{One of \code{forecast} for visualization of quantiles of the
marginal forecast distribution (i.e., the usual fanchart), or \code{paths} to
visualize a few sample paths from the joint forecast distribution}

\item{date}{Optional additional vector with dates in format that can be cast
to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create x-axis}

\item{date_future}{Optional additional vector with dates in format that can
be cast to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create
x-axis for forecast path}

\item{show_params}{Logical; if \code{TRUE} (default) then fitted params will be
displayed using \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}

\item{n}{Number of paths to add to plot, a small number is recommended to be
able to see the individual paths; positive scalar integer; used
when \code{method} is \code{"paths"}}

\item{alpha}{The transparency parameter used when adding the paths to the
plot, provided to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} and \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}; used
when \code{method} is \code{"paths"}}
}
\description{
Use \code{ggplot2} to visualize the marginal forecast quantiles, or a few sample
paths of a \code{threedx} forecast object of class \code{threedx_paths}
}
\details{
Note: This function will use \code{\link[base:sample]{base::sample()}} to randomly select paths that
are added to the plot. Set a seed if you require reproducibility.
}
