% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{draw_bootstrap}
\alias{draw_bootstrap}
\title{Draw innovations by bootstrapping from unweighted residual errors}
\usage{
draw_bootstrap(n, errors, ...)
}
\arguments{
\item{n}{The number of innovations to draw}

\item{errors}{The residual errors that are used to define the distribution
from which the innovations are drawn}

\item{...}{Additional arguments passed from \code{\link[=predict.threedx]{predict.threedx()}}, ignored}
}
\value{
A vector of same type as \code{errors} and of length \code{n}
}
\description{
Draw innovations by bootstrapping from unweighted residual errors
}
\examples{
predict(
  object = model,
  horizon = 12L,
  n_samples = 1000L,
  observation_driven = FALSE,
  innovation_function = draw_bootstrap
)
}
