% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{draw_normal_with_drift}
\alias{draw_normal_with_drift}
\title{Draw i.i.d. innovations from a Normal distribution with non-zero mean}
\usage{
draw_normal_with_drift(n, errors, ...)
}
\arguments{
\item{n}{The number of innovations to draw}

\item{errors}{The residual errors that are used to define the distribution
from which the innovations are drawn}

\item{...}{Additional arguments passed from \code{\link[=predict.threedx]{predict.threedx()}}, ignored}
}
\value{
A vector of same type as \code{errors} and of length \code{n}
}
\description{
Draw i.i.d. innovations from a Normal distribution with non-zero mean
}
\examples{
predict(
  object = model,
  horizon = 12L,
  n_samples = 1000L,
  observation_driven = FALSE,
  innovation_function = draw_normal_with_drift,
)
}
