% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{draw_normal_with_zero_mean}
\alias{draw_normal_with_zero_mean}
\title{Draw i.i.d. innovations from a Normal distribution with zero mean}
\usage{
draw_normal_with_zero_mean(n, errors, ...)
}
\arguments{
\item{n}{The number of innovations to draw}

\item{errors}{The residual errors that are used to define the distribution
from which the innovations are drawn}

\item{...}{Additional arguments passed from \code{\link[=predict.threedx]{predict.threedx()}}, ignored}
}
\value{
A vector of same type as \code{errors} and of length \code{n}
}
\description{
Draw i.i.d. innovations from a Normal distribution with zero mean
}
\examples{
model <- learn_weights(
  y = 1:50,
  period_length = 12L,
  alphas_grid = list_sampled_alphas(n_target = 25),
  loss_function = loss_mae
)

forecast <- predict(
  object = model,
  horizon = 12L,
  n_samples = 1000L,
  observation_driven = FALSE,
  innovation_function = draw_normal_with_zero_mean,
)
}
