% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.threedx}
\alias{fitted.threedx}
\title{Extract fitted values from a \code{threedx} model}
\usage{
\method{fitted}{threedx}(object, ...)
}
\arguments{
\item{object}{A \code{threedx} model as returned by \code{\link[=learn_weights]{learn_weights()}}}

\item{...}{Other arguments passed to \code{fitted()}, ignored}
}
\value{
A numeric vector of same length as \code{object$y}
}
\description{
Returns the fitted one-step-ahead predictions for the training data of the
\code{threedx} model. At least the first value will be missing.
}
