% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_alphas.R
\name{list_sampled_alphas}
\alias{list_sampled_alphas}
\title{Generate a list of possible alpha values that are evaluated during training}
\usage{
list_sampled_alphas(
  n_target = 100,
  alpha_lower = 0,
  alpha_upper = 1,
  alpha_seasonal_lower = 0,
  alpha_seasonal_upper = 1,
  alpha_seasonal_decay_lower = 0,
  alpha_seasonal_decay_upper = 1,
  oversample_lower = 0.05,
  oversample_upper = 0.05,
  include_edge_cases = TRUE,
  seed = NULL
)
}
\description{
Generate a list of possible alpha values that are evaluated during training
}
\examples{
alphas_grid <- list_sampled_alphas(n_target = 250L)

# ggplot2::ggplot(
#   alphas_list_to_data_frame(alphas_grid),
#   ggplot2::aes(x = alpha, y = alpha_seasonal, fill = alpha_seasonal_decay)
# ) +
#   ggplot2::geom_point(pch = 21, color = "white") +
#   ggplot2::scale_fill_gradient2(midpoint = 0.5) +
#   ggplot2::labs(title = "Sampled Grid of Parameters to Evaluate")

}
