% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{loss_mae_with_observation_weight}
\alias{loss_mae_with_observation_weight}
\title{Observation-weighted mean absolute error loss function}
\usage{
loss_mae_with_observation_weight(y_hat, y, ...)
}
\arguments{
\item{y_hat}{A numeric vector representing predictions}

\item{y}{A numeric vector representing observations}

\item{...}{Additional arguments passed from other functions; ignored}
}
\value{
A scalar value
}
\description{
Observation-weighted mean absolute error loss function
}
\references{
Suman Ravuri et al. (2021).
\emph{Skilful precipitation nowcasting using deep generative models of radar}.
\url{https://www.nature.com/articles/s41586-021-03854-z}
}
