% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{loss_rmse}
\alias{loss_rmse}
\title{Root mean-squared error loss function}
\usage{
loss_rmse(y_hat, y, ...)
}
\arguments{
\item{y_hat}{A numeric vector representing predictions}

\item{y}{A numeric vector representing observations}

\item{...}{Additional arguments passed from other functions; ignored}
}
\value{
A scalar value
}
\description{
Root mean-squared error loss function
}
