% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{loss_rmse_ignoring_bias}
\alias{loss_rmse_ignoring_bias}
\title{Root-mean-square error loss function ignoring bias}
\usage{
loss_rmse_ignoring_bias(y_hat, y, ...)
}
\arguments{
\item{y_hat}{A numeric vector representing predictions}

\item{y}{A numeric vector representing observations}

\item{...}{Additional arguments passed from other functions; ignored}
}
\value{
A scalar value
}
\description{
Subtracts the mean from the observed residuals before computing their
root-mean-square error. This can be helpful when predicting time series
where a bias due to a trend component is expected and can be captured
by the innovation function choice.
}
\seealso{
\code{\link[=draw_normal_with_drift]{draw_normal_with_drift()}}, \code{\link[=draw_bootstrap]{draw_bootstrap()}},
\code{\link[=loss_mae_ignoring_bias]{loss_mae_ignoring_bias()}}
}
