% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot fitted values of an \code{threedx} model}
\usage{
plot_fitted(object, date = NULL, show_params = TRUE)
}
\arguments{
\item{object}{Fitted model object returned by \code{\link[=learn_weights]{learn_weights()}}}

\item{date}{Optional additional vector with dates in format that can be cast
to \code{YYYY-MM-DD} with same length as \code{object$y}, used to create x-axis}

\item{show_params}{Logical; if \code{TRUE} (default) then fitted parameters will
be displayed using \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}
}
\description{
This function requires the \link[ggplot2:ggplot2-package]{ggplot2}. Whether its
namespace is available will be checked when the function is run. \code{ggplot2} is
only suggested, not a default import.
}
\examples{
set.seed(4278)
y <- rpois(n = 55, lambda = pmax(0.1, 1 + 10 * sinpi(1:55 / 6)))

model <- learn_weights(
  y = y,
  alphas_grid = list_sampled_alphas(n_target = 50L),
  period_length = 12L,
  loss_function = loss_mae
)
threedx:::plot_fitted(object = model)

}
\keyword{internal}
