% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{plot_paths}
\alias{plot_paths}
\title{Plot a few forecast sample paths of a \code{threedx} model}
\usage{
plot_paths(
  object,
  date = NULL,
  date_future = NULL,
  n = 5,
  alpha = 0.75,
  show_params = TRUE
)
}
\arguments{
\item{object}{An object of class \code{threedx_paths} as returned by
\code{predict.threedx()}}

\item{date}{Optional additional vector with dates in format that can be cast
to \code{YYYY-MM-DD}, used to create x-axis}

\item{date_future}{Optional additional vector with dates in format that can
be cast to \code{YYYY-MM-DD}, used to create
x-axis for forecast paths}

\item{n}{Number of paths to add to plot, a small number is recommended to be
able to see the individual paths; positive scalar integer}

\item{alpha}{The transparency parameter used when adding the paths to the
plot, provided to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} and \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}}

\item{show_params}{Logical; if \code{TRUE} (default) then fitted params will be
displayed using \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}}
}
\description{
This function requires the \link[ggplot2:ggplot2-package]{ggplot2}. Whether its
namespace is available will be checked when the function is run. \code{ggplot2} is
only suggested, not a default import.
}
\details{
Note: This function will use \code{\link[base:sample]{base::sample()}} to randomly select paths that
are added to the plot. Set a seed if you require reproducibility.
}
\examples{
set.seed(4278)
y <- rpois(n = 55, lambda = pmax(0.1, 1 + 10 * sinpi(1:55 / 6)))

model <- learn_weights(
  y = y,
  alphas_grid = list_sampled_alphas(n_target = 50L),
  period_length = 12L,
  loss_function = loss_mae
)

paths <- predict(
  object = model,
  horizon = 12,
  n_samples = 2500L,
  observation_driven = TRUE
)

threedx:::plot_paths(object = paths, n = 3)

}
\keyword{internal}
