% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_weights.R
\name{predict_one_step_ahead_with_grid}
\alias{predict_one_step_ahead_with_grid}
\title{Predict one-step ahead for each of the alpha specifications}
\usage{
predict_one_step_ahead_with_grid(y, alphas_grid, n, period_length)
}
\arguments{
\item{y}{The time series vector of length \code{n}}

\item{alphas_grid}{The list of possible parameter combinations}

\item{n}{The number of observations in the time series}

\item{period_length}{The length of the seasonal period}
}
\value{
A numeric vector of length \code{length(alphas_grid)}
}
\description{
Generates a matrix of weights in which each column contains the weights for
one of the possible paramater values that are being evaluated during
training. Then derives one-step-ahead predictions for each set of weights
by matrix multiplication with the time series vector.
}
\keyword{internal}
