% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_with_state}
\alias{predict_with_state}
\title{Generate sample paths using latent state}
\usage{
predict_with_state(y_m, object, horizon, n_samples)
}
\description{
Generate sample paths using latent state
}
\keyword{internal}
