% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.threedx}
\alias{residuals.threedx}
\title{Extract residuals from a \code{threedx} model}
\usage{
\method{residuals}{threedx}(object, ...)
}
\arguments{
\item{object}{A \code{threedx} model as returned by \code{\link[=learn_weights]{learn_weights()}}}

\item{...}{Other arguments passed to \code{residuals()}, ignored}
}
\value{
A numeric vector of same length as \code{object$y}
}
\description{
Returns the residuals from one-step-ahead predictions for the training data
of the \code{threedx} model. At least the first value will be missing.
}
