% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedx-package.R
\docType{package}
\name{threedx-package}
\alias{threedx}
\alias{threedx-package}
\title{threedx: Three-Dimensional Exponential Smoothing Forecasts}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timradtke/threedx}
  \item Report bugs at \url{https://github.com/timradtke/threedx/issues}
}

}
\author{
\strong{Maintainer}: Tim Radtke \email{tim@timradtke.net}

}
\keyword{internal}
