% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_non_negative_with_identical_mean.R
\name{to_non_negative_with_identical_mean}
\alias{to_non_negative_with_identical_mean}
\title{Turn values non-negative while preserving their sample mean}
\usage{
to_non_negative_with_identical_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Numeric vector of length equal to \code{x}
}
\description{
Postprocess function turning forecast samples non-negative while keeping the
mean at the same value as it was before through scaling of the samples
(unless the sample mean was originally negative).
}
\details{
While the identical sample mean can be enforced, this will generally decrease
the sample variance as this is the only way to maintain the mean while
turning negative samples to zero to enforce the non-negativity constraint.

The returned samples will always be non-negative. The sample mean will not be
identical if it cannot be achieved given the non-negative samples.
}
\examples{
to_non_negative_with_identical_mean(x = c(-1, 0, 1, 2))

x <- rnorm(100, 0.5)
summary(x)
summary(to_non_negative_with_identical_mean(x))

x <- rnorm(100, -1)
summary(x)
# can't keep identical mean as original mean is negative
summary(to_non_negative_with_identical_mean(x))

}
\seealso{
\code{\link[=predict.threedx]{predict.threedx()}}, \code{\link[=to_moment_matched_nbinom]{to_moment_matched_nbinom()}}
}
