% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_exponential.R
\name{weights_exponential}
\alias{weights_exponential}
\title{Derive exponential weights}
\usage{
weights_exponential(alpha, n)
}
\description{
A larger value of \code{alpha} will assign larger weights to more recent
observations.
}
