% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_exponential.R
\name{weights_exponential_vec}
\alias{weights_exponential_vec}
\title{Vectorized \code{weights_exponential()}}
\usage{
weights_exponential_vec(alphas, n)
}
\description{
Vectorized \code{weights_exponential()}
}
\examples{
t(vapply(
  X = c(1, 0.2, 0.5, 0.8, 0),
  FUN = weights_exponential,
  n = 50,
  FUN.VALUE = numeric(50),
  USE.NAMES = FALSE
))


}
\keyword{internal}
