% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_seasonal.R
\name{weights_seasonal}
\alias{weights_seasonal}
\title{Derive seasonal exponential weights}
\usage{
weights_seasonal(alpha_seasonal, n, season_length)
}
\description{
Derive seasonal exponential weights
}
