% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_seasonal.R
\name{weights_seasonal_decay}
\alias{weights_seasonal_decay}
\title{Derive weights with seasonal exponential decay}
\usage{
weights_seasonal_decay(alpha_seasonal_decay, n, season_length)
}
\description{
Derive weights with seasonal exponential decay
}
