% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_seasonal.R
\name{weights_seasonal_decay_vec}
\alias{weights_seasonal_decay_vec}
\title{Vectorized \code{weights_seasonal_decay()}}
\usage{
weights_seasonal_decay_vec(alphas_seasonal_decay, n, period_length)
}
\description{
Vectorized \code{weights_seasonal_decay()}
}
\examples{
t(vapply(
  X = c(1, 0.2, 0.5, 0.8, 0),
  FUN = weights_seasonal_decay,
  n = 50,
  period_length = 12,
  FUN.VALUE = numeric(50),
  USE.NAMES = FALSE
))

threedx:::weights_seasonal_decay_vec(
  alphas_seasonal_decay = c(0.2, 0.5, 0.8), n = 50, period_length = 12L
)
}
\keyword{internal}
