% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_threedx.R
\name{weights_threedx_vec}
\alias{weights_threedx_vec}
\title{Vectorized \code{weights_threedx()}}
\usage{
weights_threedx_vec(
  alphas,
  alphas_seasonal,
  alphas_seasonal_decay,
  n,
  period_length
)
}
\description{
Vectorized \code{weights_threedx()}
}
\examples{
threedx:::weights_threedx_vec(
  alphas = c(0, 0.2, 0.5, 0.8, 1),
  alphas_seasonal = c(0.1, 0.3, 0.55, 1, 0.25),
  alphas_seasonal_decay = c(0.11, 0.01, 0.02, 0.05, 0.1),
  n = 50,
  period_length = 7
)
}
\keyword{internal}
