% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_alphas.R
\name{alphas_list_to_data_frame}
\alias{alphas_list_to_data_frame}
\title{Cast a list of alphas as data frame}
\usage{
alphas_list_to_data_frame(alphas_grid)
}
\description{
Cast a list of alphas as data frame
}
\examples{
alphas_list_to_data_frame(list_sampled_alphas(n_target = 10L))

}
