% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_weights.R
\name{learn_weights}
\alias{learn_weights}
\title{Fit a 3DX model to a time series}
\usage{
learn_weights(y, period_length, alphas_grid, loss_function)
}
\arguments{
\item{y}{The time series to be forecasted}

\item{period_length}{The presumed length of a seasonal period for \code{y}}

\item{alphas_grid}{A data frame of possible parameter combinations to
generate the weights of the final model. The optimal parameter set will be
chosen based on the minimization of \code{loss_function}. The expected columns
are numeric and called \code{alpha}, \code{alpha_seasonal}, \code{alpha_seasonal_decay}.
At least one row must be provided.
This list can be generated via \code{\link[=list_sampled_alphas]{list_sampled_alphas()}} or
\code{\link[=list_edge_alphas]{list_edge_alphas()}}, for example.}

\item{loss_function}{A function with first argument \code{y_hat} and optionally
more arguments. Usually, to compute a loss, at least \code{y} is required to
compute errors. Must be able to handle additional parameters via \code{...} to
allow for potential future changes in the set of arguments passed to
\code{loss_function} by \code{\link[=learn_weights]{learn_weights()}}. For examples, see \code{\link[=loss_mae]{loss_mae()}} or
\code{\link[=loss_mae_with_observation_weight]{loss_mae_with_observation_weight()}}.
The arguments \code{y_hat} and \code{y} passed by \code{\link[=learn_weights]{learn_weights()}} are numeric
vectors of equal length.
The provided \code{loss_function} must return a numeric scalar value.}
}
\value{
A fitted model object of class \code{threedx}
}
\description{
Fit a 3DX model to a time series
}
\examples{
set.seed(9284)
y <- stats::rpois(n = 55, lambda = pmax(0.1, 1 + 10 * sinpi((5 + 1:55 )/ 6)))

model <- learn_weights(
  y = y,
  alphas_grid = list_sampled_alphas(
    n_target = 1000L,
    include_edge_cases = TRUE
  ),
  period_length = 12L,
  loss_function = loss_mae
)

if (require("ggplot2")) {
  autoplot(model)
}
}
