% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_alphas.R
\name{list_edge_alphas}
\alias{list_edge_alphas}
\title{Generate a list of possible alpha values that are evaluated during training}
\usage{
list_edge_alphas()
}
\description{
Generate a list of possible alpha values that are evaluated during training
}
