% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{loss_mae_with_observation_weight}
\alias{loss_mae_with_observation_weight}
\title{Observation-weighted mean-absolute error loss function}
\usage{
loss_mae_with_observation_weight(y_hat, y, ...)
}
\description{
Observation-weighted mean-absolute error loss function
}
