% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.threedx}
\alias{predict.threedx}
\title{Draw forecast sample paths from a fitted 3DX model}
\usage{
\method{predict}{threedx}(object, horizon, n_samples, observation_driven, ...)
}
\description{
Draw forecast sample paths from a fitted 3DX model
}
