% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_with_observations}
\alias{predict_with_observations}
\title{Generate sample paths using observations}
\usage{
predict_with_observations(y_m, object, horizon, n_samples)
}
\description{
Generate sample paths using observations
}
\keyword{internal}
