% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedx-package.R
\docType{package}
\name{threedx-package}
\alias{threedx}
\alias{threedx-package}
\title{threedx: Three-Dimensional Exponential Smoothing Forecasts}
\description{
Use 3DX to generate interpretable probabilistic forecasts by purely weighting values from the observed time series. Unique about 3DX is that it can be used to derive forecasts not only based on a “latent state” but also observation-driven, drawing future realizations purely from observed values. This can be effective for count or intermittent series. The 3DX sampling distribution is a combination of three components: The first weighs observations exponentially, the second weighs observations within each seasonal period exponentially, the third weighs the seasonal periods over time exponentially.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timradtke/threedx}
  \item Report bugs at \url{https://github.com/timradtke/threedx/issues}
}

}
\author{
\strong{Maintainer}: Tim Radtke \email{tim@timradtke.net}

}
\keyword{internal}
