% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_exponential.R
\name{weights_exponential}
\alias{weights_exponential}
\title{Derive exponential weights}
\usage{
weights_exponential(alpha, n)
}
\arguments{
\item{alpha}{A value between 0 and 1 that determines how quick the
exponential decay in the weights is}

\item{n}{The number of weights to create; this is usually equal to the
number of observations in a time series}
}
\value{
A monotonically increasing numeric vector of \code{n} values between 0
and 1 that sum up to 1
}
\description{
A larger value of \code{alpha} will assign larger weights to more recent
observations.
}
\examples{
weights_exponential(alpha = 0.25, n = 5)
weights_exponential(alpha = 1, n = 7)
weights_exponential(alpha = 0, n = 4)

# zooming into an exponential series again gives an exponential series
identical(
  weights_exponential(alpha = 0.75, n = 5)[1:4] /
    sum(weights_exponential(alpha = 0.75, n = 5)[1:4]),
  weights_exponential(alpha = 0.75, n = 4)
)

}
\seealso{
\code{\link[=weights_seasonal]{weights_seasonal()}}, \code{\link[=weights_seasonal_decay]{weights_seasonal_decay()}},
\code{\link[=weights_threedx]{weights_threedx()}}
}
