% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_seasonal.R
\name{weights_seasonal}
\alias{weights_seasonal}
\title{Derive seasonal exponential weights}
\usage{
weights_seasonal(alpha_seasonal, n, period_length)
}
\arguments{
\item{alpha_seasonal}{A value between 0 and 1 that determines how quick the
exponential decay in the weights is}

\item{n}{The number of weights to create; this is usually equal to the
number of observations in a time series}

\item{period_length}{The length of the seasonal period to be modeled}
}
\value{
A numeric vector of \code{n} values between 0 and 1 that sum up to 1
}
\description{
Assigns the largest weight to the same season of the period with length
\code{period_length}, and exponentially decaying weights to the neighboring
seasons. The most recent value in the resulting weights vector will be
assigned the second-highest weight as it is the direct neighbor of the
current season. See examples for illustrations of this.
}
\examples{
round(
  weights_seasonal(alpha_seasonal = 0.25, n = 8, period_length = 7L),
  3
)

round(
  weights_seasonal(alpha_seasonal = 0.75, n = 27, period_length = 12L),
  3
)

# "seasonal mean" weights
weights_seasonal(alpha_seasonal = 1, n = 30, period_length = 7L)

# mean weights
weights_seasonal(alpha_seasonal = 0, n = 30, period_length = 7L)

}
\seealso{
\code{\link[=weights_seasonal_decay]{weights_seasonal_decay()}}, \code{\link[=weights_exponential]{weights_exponential()}},
\code{\link[=weights_threedx]{weights_threedx()}}
}
